Fix up circular dependencies
- Move mako_criteria_spec to types.h
- Use more forward declarations where full types are unnecessary
- Explicitly include wayland-client.h in criteria.h
