+++ include/config.h
@@ -4,8 +4,6 @@
 #include <stdbool.h>
 #include <stdint.h>
 #include <wayland-client.h>
-
-#include "criteria.h"
 #include "types.h"
 #include "wlr-layer-shell-unstable-v1-client-protocol.h"
